requirements.txt
altair==5.1.2
attrs==23.1.0
bitarray==2.8.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
imageio==2.32.0
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
networkx==3.2.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.0.1
protobuf==4.25.0
pyarrow==14.0.0
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
scikit-image==0.22.0
scipy==1.11.3
six==1.16.0
smmap==5.0.1
streamlit==1.28.1
tenacity==8.2.3
tifffile==2023.9.26
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
